Move validate from Profiles.validate to validate.js
