+++ src/profiles.js
@@ -4,8 +4,7 @@ import path from 'path'
 
 import Utils from './utils'
 
-const PROFILES_CACHED = {}
-  , DEFAULT_REMOTE_PATH = 'http://schemas.datapackages.org/registry.csv'
+const DEFAULT_REMOTE_PATH = 'http://schemas.datapackages.org/registry.csv'
 let DEFAULT_LOCAL_PATH
 if (!Utils.isBrowser) {
   DEFAULT_LOCAL_PATH = path.join(__dirname, '..', 'schemas', 'registry.csv')
@@ -159,37 +158,6 @@ class Profiles {
   }
 }
 
-/**
- * Standalone function for validating datapackage descriptor against a profile.
- * It encapsulates the Profiles class and exposes only validation. Profile
- * promises are cached and the class will not be initialized on every call.
- *
- * @param descriptor
- * @param profile
- * @param remoteProfiles
- * @return {Promise} Resolves `true` or Array of errors.
- */
-function validate(descriptor, profile = 'base',
-  remoteProfiles = 'false') {
-  const remoteString = remoteProfiles.toString()
-
-  if (PROFILES_CACHED[remoteString]) {
-    return new Promise(resolve => {
-      PROFILES_CACHED[remoteString].then(profiles => {
-        resolve(profiles.validate(descriptor, profile))
-      })
-    })
-  }
-
-  PROFILES_CACHED[remoteString] = new Profiles(remoteProfiles)
-
-  return new Promise(resolve => {
-    PROFILES_CACHED[remoteString].then(profiles => {
-      resolve(profiles.validate(descriptor, profile))
-    })
-  })
-}
-
-export { Profiles, validate }
+export default Profiles
 
 /* eslint arrow-body-style: "off" */
